% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_models.R
\name{kernel_fda}
\alias{kernel_fda}
\title{Flexible Discriminant Analysis with user-defined kernel}
\usage{
kernel_fda(formula, data, kerneldot = NULL, lambda = 1, ...)
}
\arguments{
\item{formula}{formula of the form ‘y~x’ it describes the response and the
predictors.  The formula can be more complicated, such as
‘y~log(x)+z’ etc (see ‘formula’ for more details).  The
response should be a factor representing the response
variable, or any vector that can be coerced to such (such as
a logical variable).}

\item{data}{data.frame containing predictor and response data.  The predictor
data should be centered and scaled (normalized).}

\item{kerneldot}{method from \code{\link[kernlab]{dots}} methods or
user-defined kernel function with the same signature.  Default is
linear kernel.}

\item{lambda}{numeric regularization coefficient.  Default is 1.}

\item{...}{additional parameters passed to kernel.dot}
}
\value{
\code{mda} \code{fda} object.
}
\description{
Builds a regularized flexible discriminant model from user supplied
data, kernel function, and regularization parameter.
}
\examples{
data(iris)

# Normalize data
means <- apply(iris[,1:4],2,mean)
stddevs <- apply(iris[,1:4],2,sd)
iris_normalized <- iris
for (col in 1:4) {
  iris_normalized[,col] <- (iris[,col] - means[col])/stddevs[col]
}

model <- kernel_fda(
  Species ~ .,
  iris_normalized,
  kerneldot = kernlab::rbfdot,
  lambda = 1,
  sigma = 0.1
)
}
\references{
Hastie, T., Tibshirani, R., & Friedman, J. H. (2009),
\emph{The elements of statistical learning: data mining, inference,
and prediction,} 2nd ed. New York: Springer, pp 106--119 and pp 440--455.
\url{https://hastie.su.domains/ElemStatLearn/printings/ESLII_print12_toc.pdf}
}
\seealso{
\code{\link[mda]{fda}}, \code{\link[kernlab]{dots}},
\code{\link{linear_fda}},
\url{https://hastie.su.domains/ElemStatLearn/printings/ESLII_print12_toc.pdf}
}
