% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_models.R
\name{kernel_reg_model}
\alias{kernel_reg_model}
\title{Kernel Regression}
\usage{
kernel_reg_model(x, y, w, kernel_function, lambda = 1)
}
\arguments{
\item{x}{matrix of predictor data.}

\item{y}{vector or matrix of response data.}

\item{w}{numeric not used.  Intended to be weights to be applied to training
data.}

\item{kernel_function}{function to compute the kernel.}

\item{lambda}{numeric regularization coefficient.  Default is 1.}
}
\value{
function that takes to matrices and returns the
dot product
}
\description{
Construct a linear regression model using the user-supplied
kernel function.  This function is called by \code{kernel_fda} and
is not intended to be used outside of that context.
}
\seealso{
\code{\link{kernel_fda}}
}
