% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_models.R
\name{linear_fda}
\alias{linear_fda}
\title{Flexible Discriminant Analysis with linear regression model}
\usage{
linear_fda(formula, data, lambda = 1)
}
\arguments{
\item{formula}{formula of the form ‘y~x’ it describes the response and the
predictors.  The formula can be more complicated, such as
‘y~log(x)+z’ etc (see ‘formula’ for more details).  The
response should be a factor representing the response
variable, or any vector that can be coerced to such (such as
a logical variable).}

\item{data}{data.frame containing predictor and response data.  The
predictor data should be centered and scaled (normalized).}

\item{lambda}{numeric regularization coefficient.  Default is 1.}
}
\value{
\code{mda} \code{fda} object.
}
\description{
Builds a regularized flexible discriminant model from user supplied
data and L2-regularization parameter, using linear regression model.
This method is useful for quickly comparing more complicated models
to a linear model and for debugging more complicated methods.  By
itself, it does not provide functionality beyond what is already
provided in the \code{mda} package, and might be less efficient
than equivalent methods there.
}
\examples{
data(iris)

# Normalize data
means <- apply(iris[,1:4],2,mean)
stddevs <- apply(iris[,1:4],2,sd)
iris_normalized <- iris
for (col in 1:4) {
  iris_normalized[,col] <- (iris[,col] - means[col])/stddevs[col]
}

model <- linear_fda(
  Species ~ .,
  iris_normalized,
  lambda = 1
)
}
\seealso{
\code{\link[mda]{fda}}, \code{\link[kernlab]{dots}},
\code{\link{kernel_fda}},
\url{https://hastie.su.domains/ElemStatLearn/printings/ESLII_print12_toc.pdf}
}
