% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_models.R
\name{linear_reg_model}
\alias{linear_reg_model}
\title{Linear Regression}
\usage{
linear_reg_model(x, y, w, lambda = 1)
}
\arguments{
\item{x}{matrix of predictor data.}

\item{y}{vector or matrix of response data.}

\item{w}{numeric not used.  Intended to be weights to be applied to training
data.}

\item{lambda}{numeric regularization coefficient.  Default is 1.}
}
\value{
function that takes to matrices and returns the
dot product
}
\description{
Construct a L2-regularized linear regression model.
This function is called by \code{kernel_fda} and
is not intended to be used outside of that context.
}
\seealso{
\code{\link{linear_fda}}
}
