% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_models.R
\name{lineardot}
\alias{lineardot}
\title{Linear kernel method}
\usage{
lineardot(add_ones = FALSE)
}
\arguments{
\item{add_ones}{logical if TRUE will add an intercept column
(i.e. a column of 1's) to input matricies}
}
\value{
function that takes to matrices and returns the
dot product
}
\description{
Generate linear kernel matrix (i.e., dot product).  This
function is meant to be used to build a kernel_model object,
(from the \code{kernel_fda} method) and not meant to be used explicity.
}
\seealso{
\code{\link{kernel_fda}}, \code{\link[kernlab]{dots}}
}
