% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_kda}
\alias{plot_kda}
\title{Plot a flexible linear discriminant model}
\usage{
plot_kda(model_obj, newdata, xcoord = 1, ycoord = 2)
}
\arguments{
\item{model_obj}{\code{mda} \code{fda} object.}

\item{newdata}{data frame new data to be plotted.  Must have same names as
data used for fitting.  If omitted, training data will be plotted.}

\item{xcoord}{integer coordinate to use on x axis.}

\item{ycoord}{integer coordinate to use on y axis.}
}
\value{
\code{ggplot2} \code{ggplot} object
}
\description{
Plots centroids and cannonical variates in specified coordinates.
The coordinate values must be strictly less than the number of classes.
(See p. 445, Hastie, et al.)
in the model.  The first coordinate correspondes to the largest eigenvalue.
}
\examples{
data(iris)

# Normalize data
means <- apply(iris[,1:4], 2, mean)
stddevs <- apply(iris[,1:4], 2, sd)
iris_normalized <- iris
for (col in 1:4) {
  iris_normalized[, col] <- (iris[, col] - means[col]) /
    stddevs[col]
}

model <- kernel_fda(
  Species ~ .,
  iris_normalized,
  kerneldot = kernlab::rbfdot,
  lambda = 1,
  sigma = 0.1
)

plot(model)
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link{kernel_fda}},
\code{\link{linear_fda}}
}
