% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.linear_regression}
\alias{predict.linear_regression}
\title{Predict from linear regression model}
\usage{
\method{predict}{linear_regression}(object, newdata, ...)
}
\arguments{
\item{object}{linear regression model object.}

\item{newdata}{matrix containing new predictor data.  If not provided,
training data will be used.}

\item{...}{additional arguments to pass to kernel_regression (not used).}
}
\value{
matrix of predicted values
}
\description{
Estimate response values using a linear regression model.  This method
gets called implicitly in the \code{predict.fda} method when the
fda model is generated by the \code{linear_fda} method. It is not
intended to be used outside of that context.
}
\seealso{
\code{\link{linear_reg_model}}
}
