% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{test_kda}
\alias{test_kda}
\title{Test model performance}
\usage{
test_kda(model_obj, newdata, print_results = TRUE)
}
\arguments{
\item{model_obj}{\code{mda} \code{fda} model.}

\item{newdata}{data frame providing data to investigate.  Must include
all of the predictor variables and the response variable.  If omitted,
training data will be used.}

\item{print_results}{logical.  If TRUE, results will be outputted to
stdout}
}
\value{
numeric correct classification rate.
}
\description{
Compute the correct classification rate and misclassification rate
for a fit fda model.
}
\examples{
data(iris)

# Separate into training & test
data_permutation_order <- sample(nrow(iris))
train_cutoff <- round(0.75 * nrow(iris))
train_indices <- data_permutation_order[1:train_cutoff]
test_indices <- data_permutation_order[(train_cutoff + 1):nrow(iris)]
train_data_unnorm <- iris[train_indices, ]
test_data_unnorm <- iris[test_indices, ]

# Normalize data
means <- apply(train_data_unnorm[,1:4],2,mean)
stddevs <- apply(train_data_unnorm[,1:4],2,sd)
train_data_norm <- train_data_unnorm
test_data_norm <- test_data_unnorm

for (col in 1:4) {
  train_data_norm[, col] <- (train_data_unnorm[, col] -
    means[col]) / stddevs[col]
  test_data_norm[, col] <- (test_data_unnorm[, col] -
    means[col]) / stddevs[col]
}

model <- kernel_fda(
  Species ~ .,
  train_data_norm,
  kerneldot = kernlab::rbfdot,
  lambda = 1,
  sigma = 0.1
)

test_kda(model)
test_kda(model,test_data_norm)
}
